fix(web): move conversation chart helpers out of component module
Keep KeyedConversationTable hot-refresh compatible by moving shared prompt cache chart helpers into a non-component module.
